Update typescript ESlint rules to work with JSX and Preact
## Summary
ESlint was throwing up a few common errors for every Preact .tsx file.
This includes not recognizing Preact's 'h' import as being used, and not
recognizing any components as being used, either. In addition, we need
to set the return type of each render() function as JSX.Element.

## Motivation
Enable the use of ESlint without having to sort out the things we can't
possibly fix.