Code-split preact into a seperate bundle
## Summary
Specific parts of Preact ('h', 'Component', and 'render') have been
split out into a new bundle called 'vendor.[hash].js'. This file should
only include code which will be imported for every page on the website,
which is Preact for now.

## Motivation
Since Preact needs to be included on every single page of the website,
it would add ~3 KB to the bundle size of every single page. Instead, we
can split it out and then aggressively cache the resulting file, making
the empty bundle size tiny.